Spread implementation over separate files
