+++ Sources/Player/Types/MediaSelector.swift
@@ -6,102 +6,6 @@
 import AVFoundation
 import MediaAccessibility
 
-public enum MediaSelectionOption: Hashable {
-    case automatic
-    case disabled
-    case enabled(AVMediaSelectionOption)
-
-    public var displayName: String {
-        switch self {
-        case .automatic:
-            return "Auto (Recommended)"
-        case .disabled:
-            return "Off"
-        case let .enabled(option):
-            return option.displayName
-        }
-    }
-}
-
-protocol MediaSelectionGroup {
-    var options: [MediaSelectionOption] { get }
-
-    init(group: AVMediaSelectionGroup)
-
-    func selectedMediaOption(in selection: AVMediaSelection) -> MediaSelectionOption
-    func select(mediaOption: MediaSelectionOption, in item: AVPlayerItem)
-}
-
-struct AudibleGroup: MediaSelectionGroup {
-    let group: AVMediaSelectionGroup
-
-    var options: [MediaSelectionOption] {
-        guard group.options.count > 1 else { return [] }
-        return group.options.map { .enabled($0) }
-    }
-
-    func selectedMediaOption(in selection: AVMediaSelection) -> MediaSelectionOption {
-        if let option = selection.selectedMediaOption(in: group) {
-            return .enabled(option)
-        }
-        else {
-            return .disabled
-        }
-    }
-
-    func select(mediaOption: MediaSelectionOption, in item: AVPlayerItem) {
-        switch mediaOption {
-        case let .enabled(option):
-            item.select(option, in: group)
-        default:
-            item.select(nil, in: group)
-        }
-    }
-}
-
-struct LegibleGroup: MediaSelectionGroup {
-    let group: AVMediaSelectionGroup
-
-    var options: [MediaSelectionOption] {
-        var options: [MediaSelectionOption] = [.automatic, .disabled]
-        options.append(contentsOf: group.options.map { .enabled($0) })
-        return options
-    }
-
-    func selectedMediaOption(in selection: AVMediaSelection) -> MediaSelectionOption {
-        switch MACaptionAppearanceGetDisplayType(.user) {
-        case .alwaysOn:
-            if let option = selection.selectedMediaOption(in: group) {
-                return .enabled(option)
-            }
-            else {
-                return .disabled
-            }
-        case .automatic:
-            return .automatic
-        default:
-            return .disabled
-        }
-    }
-
-    func select(mediaOption: MediaSelectionOption, in item: AVPlayerItem) {
-        switch mediaOption {
-        case .disabled:
-            MACaptionAppearanceSetDisplayType(.user, .forcedOnly)
-            item.select(nil, in: group)
-        case .automatic:
-            MACaptionAppearanceSetDisplayType(.user, .automatic)
-            item.selectMediaOptionAutomatically(in: group)
-        case let .enabled(option):
-            MACaptionAppearanceSetDisplayType(.user, .alwaysOn)
-            if let languageCode = option.locale?.language.languageCode {
-                MACaptionAppearanceAddSelectedLanguage(.user, languageCode.identifier as CFString)
-            }
-            item.select(option, in: group)
-        }
-    }
-}
-
 /// Manages the available media selections as well as the currently selected one.
 struct MediaSelector {
     static var empty: Self {
